GM/T 0001-2012 - Zu Chongzhi Stream Cipher
Zu Chongzhi (祖沖之, 429 – 500 CE) was a Chinese astronomer, mathematician, writer, politician and inventor.

He calculated Pi accurately to 7 decimal places – a record which was not surpassed until 800 years later. To do this, he approximated a circle with a 24,576-sided polygon.

Zu also discovered the formula 4/3πr3 for the volume of a sphere. His precise astronomical observations allowed him to create a new, more accurate calendar and to predict solar eclipses. He also calculated that Jupiter takes almost 12 years to orbit the sun.